requirements.txt
black==21.9b0
boto3==1.11.17
botocore==1.14.17
click==8.0.2
docutils==0.15.2
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pymongo==3.10.1
python-dateutil==2.8.2
python-dotenv==0.19.1
PyYAML==5.4.1
regex==2021.10.8
s3transfer==0.3.7
six==1.16.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.25.11
